dashboard/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
aiohttp==3.9.1
pydantic==2.5.0
